refactor osserial and libftdiserial
