+++ src/utils/os_serial_ioctl.cpp
@@ -1,4 +1,4 @@
-#include "serial.h"
+#include "os_serial.h"
 
 // ioctl headers conflict with standard posix headers for serial port, ioctl needed for custom
 // baudrate, move this function to another compilation unit from serial.cpp
@@ -23,11 +23,9 @@
 
 #endif
 
-// For platform-specific definition code
-#define Serial OSSerial
 
 #if defined(_WIN32)
-int Serial::set_custom_baudrate (int baudrate)
+int OSSerial::set_custom_baudrate (int baudrate)
 {
     DCB dcb_serial_params = {0};
     dcb_serial_params.DCBlength = sizeof (dcb_serial_params);
@@ -48,12 +46,12 @@ int Serial::set_custom_baudrate (int baudrate)
 
 #elif defined(__linux__) && !defined(__ANDROID__)
 #ifdef NO_IOCTL_HEADERS
-int Serial::set_custom_baudrate (int baudrate)
+int OSSerial::set_custom_baudrate (int baudrate)
 {
     return SerialExitCodes::NO_SYSTEM_HEADERS_FOUND_ERROR;
 }
 #else
-int Serial::set_custom_baudrate (int baudrate)
+int OSSerial::set_custom_baudrate (int baudrate)
 {
     struct termios2 port_settings;
     memset (&port_settings, 0, sizeof (port_settings));
@@ -77,7 +75,7 @@ int Serial::set_custom_baudrate (int baudrate)
 #endif
 
 #elif defined(__APPLE__)
-int Serial::set_custom_baudrate (int baudrate)
+int OSSerial::set_custom_baudrate (int baudrate)
 {
     struct termios port_settings;
     memset (&port_settings, 0, sizeof (port_settings));
@@ -93,7 +91,7 @@ int Serial::set_custom_baudrate (int baudrate)
 }
 
 #else
-int Serial::set_custom_baudrate (int baudrate)
+int OSSerial::set_custom_baudrate (int baudrate)
 {
     return SerialExitCodes::SET_PORT_STATE_ERROR;
 }
